Memorize the set of ordinal attnos for RETURNING
This set of columns can then be passed down to table AM functions such
as tuple_fetch_row_version(). This would further enable a columnar AM to
retrieve datums for only those columns present in a RETURNING clause.

The set of ordinal attnos is captured per RangeTblEntry. This set can
only be captured early on during the analyze-and-rewrite phase as later
on, attribute numbers don't refer to the ordinal attribute numbers for a
base relation. Thus, we capture this information as a part of
transformReturningList().

Reviewed-by: Heikki Linnakangas <heikki.linnakangas@iki.fi>