Make it mandatory to enumerate all columns to project all columns
This changes the expectation that one could get away with specifying an
empty projection list to indicate that all columns need to be projected.

Also, this introduces a FIXME to indicate that we aren't able to compute
the list of returningCols successfully for views and passes in all
columns in the ExecDelete RETURNING callsite.

Reviewed-by: Heikki Linnakangas <heikki.linnakangas@iki.fi>