Pass column projection list to tuple_lock()
Tuple locks in theory don't need the entire tuple to be fetched - just
the tid should be enough. However, callsites are always followed by an
EPQ check. So for maximum efficiency, this patch:

1. Builds a bitmap of columns needed to perform an EPQ check that
follows a lock operation by looking at the EPQState's qual. TODO: move
this to planning like scanCols.

2. Builds a bitmap of columns needed to perform an ON CONLFICT DO UPDATE
operation.

3. Passes the list down to the tuple_lock() API. To the heap, the
parameter is dead but to Zedstore, in zedstoream_lock_tuple(), this list
is passed on to the table_tuple_fetch_row_version() call.

TODO:
Callsites left to handle: triggers (GetTupleForTrigger()). Here, the
tuple slot after the lock is used for more than just EPQ

Reviewed-by: Heikki Linnakangas <heikki.linnakangas@iki.fi>