Introduce notnull_attrs for RelOptInfo.
RelOptInfo.notnull_attrs records which Var is not nullable
for the given RelOptInfo after all the quals on this RelOptInfo
are executed.

For a joinrel, the length of notnull_attrs array equals the
max value in relids, the array index is varno and the value
is a Bitmapset   which presents which varattnos are not
nullable for the varno.

For baserel, to save space, notnull_attrs[0] is always used
for current RelOptInfo.

We calculate the notnull_attrs from 2 places. one is catalog.
one is the qual filters.  In implementation, we check thems
on get_relation_info & set_xxx_size. In this way, both inherit
relation and subquery are handled easily.

Upper relation likes DISTINCT, GROUP BY, SORT, UNION / UNION ALL
is not handled in this patch.