Introduce notnull_attrs for RelOptInfo.
RelOptInfo.notnull_attrs records which Vars are not nullable
for the given RelOptInfo after all the quals on this RelOptInfo
are executed.

For a joinrel, the length of notnull_attrs array equals the
(max value + 1) in relids, the array index is varno and the
value is a Bitmapset* which presents which varattnos are
not nullable for the varno.

For baserel, to save space, notnull_attrs[0] is always used.

We calculate the notnull_attrs mainly from 2 places. one is
catalog. which is set during get_relation_info. the other one
is the qual filters, I set them during set_rel_size & set_joinrel_size
to covers as many as relkind as possible. Inherit can be handled
automatically in this way. nulls introduced by outer join
is also handled during build_join_rel.

I assumed upper relation except SETOP would have the same
notnull_attrs as the input relation, like SORT / GROUP
and so on.

SubQuery is handled but SetOp upper relation is not handled so far.

XXX: There are some code like

if (!enable_geqo)
{
elog(INFO, XXXX);
}

They are just there for easily review & testing, will be removed
at last.